+++ lve_window.h
@@ -14,4 +14,4 @@ typedef struct lve_window{
 
 lve_window* lvewin_make(int w, int h, const char* name);
 void lvewin_destroy(lve_window* lvewin);
-int lvewin_shouldclose(lve_window* lvewin);
+uint8_t lvewin_shouldclose(lve_window* lvewin);
